Issue #94: Refactored utils into ODataUtils, separated functions
